Part 06 Data Seeding

+++ Talabat.Core.Domain/Entites/Products/Product.cs
@@ -9,10 +9,11 @@ namespace Talabat.Core.Domain.Entites.Products
 {
     public class Product:BaseEntity<int>
     {
+        
         public required string Name { get; set; }
         public required string Description { get; set; }
         public string? PictureUrl { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
 
         public int? BrandId { get; set; }
         public virtual ProductBrand? Brand { get; set; } // Foriegn Key --> ProductBrand Entity
